Add confirm and prompt methods + refactoring

+++ src/pages/list-edit.page.ts
@@ -1,7 +1,6 @@
 import { Type } from '@angular/core';
-import { AlertController, ModalController, Alert, Modal } from 'ionic-angular';
+import { AlertController, ModalController, Modal } from 'ionic-angular';
 import { ReorderIndexes } from 'ionic-angular/components/item/item-reorder';
-import { Dexie } from 'dexie';
 
 import { WithId } from '../models/with-id.model';
 import { Params } from './base.page';
@@ -12,13 +11,14 @@ export abstract class ListEditPage
 <T extends WithId, U extends any, V extends WithId = WithId>
 extends ListPage<T, U, V> {
   constructor(
-    protected readonly alertController: AlertController,
+    alertController: AlertController,
     modalController: ModalController,
     modalPage: Type<ItemEditModalPage<T, U>>,
     service: U,
     navParams?: Params<V>
   ) {
     super(
+      alertController,
       modalController,
       modalPage,
       service,
@@ -37,21 +37,12 @@ extends ListPage<T, U, V> {
   }
 
   remove(item: T): void {
-    const alert: Alert = this.alertController.create({
-      title: "Are you sure?",
-      buttons: [
-        { text: 'No' },
-        { text: 'Yes', handler: () => { this.delete(item); } }
-      ]
+    this.confirm(() => {
+      this.service.delete(item).then(() => this.refresh());
     });
-    alert.present();
   }
 
   reorder($event: ReorderIndexes): void {
     $event.applyTo(this.list);
-  }
-
-  protected delete(item: T): Dexie.Promise<number> {
-    return this.service.delete(item).then(() => this.refresh());
   }
 }
